fixed integration tests
