+++ python/tests/integration/memory/vector_stores/test_vector_store.py
@@ -5,20 +5,12 @@ from typing import Any
 
 import pandas as pd
 import pytest
+from mistralai_gcp import Type
 
 from semantic_kernel.connectors.memory.redis.const import RedisCollectionTypes
-from semantic_kernel.data import VectorStore, VectorStoreRecordDefinition
+from semantic_kernel.data import VectorStore
 from semantic_kernel.exceptions import MemoryConnectorConnectionException
-from tests.integration.memory.vector_stores.data_records import (
-    PANDAS_RECORD_DEFINITION,
-    PANDAS_RECORD_DEFINITION_FLAT,
-    RAW_RECORD_ARRAY,
-    RAW_RECORD_LIST,
-    TestDataModelArray,
-    TestDataModelArrayFlat,
-    TestDataModelList,
-    TestDataModelListFlat,
-)
+from tests.integration.memory.vector_stores.data_records import RAW_RECORD_ARRAY, RAW_RECORD_LIST
 from tests.integration.memory.vector_stores.vector_store_test_base import VectorStoreTestBase
 
 
@@ -30,15 +22,20 @@ class TestVectorStore(VectorStoreTestBase):
 
     @pytest.mark.asyncio
     @pytest.mark.parametrize(
-        "store_id, collection_name, collection_options, data_model_type, data_model_definition, record",
+        "store_id, collection_name, collection_options, data_model_type, data_model_definition, distance_function, "
+        "index_kind, vector_property_type, dimensions, record",
         [
             # region Redis
             pytest.param(
                 "redis",
                 "redis_json_array_data_model",
                 {"collection_type": RedisCollectionTypes.JSON},
-                TestDataModelArray,
+                "dataclass_vector_data_model_array",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_ARRAY,
                 id="redis_json_array_data_model",
             ),
@@ -46,8 +43,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "redis",
                 "redis_json_list_data_model",
                 {"collection_type": RedisCollectionTypes.JSON},
-                TestDataModelList,
+                "dataclass_vector_data_model",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="redis_json_list_data_model",
             ),
@@ -56,7 +57,11 @@ class TestVectorStore(VectorStoreTestBase):
                 "redis_json_pandas_data_model",
                 {"collection_type": RedisCollectionTypes.JSON},
                 pd.DataFrame,
-                PANDAS_RECORD_DEFINITION,
+                "data_model_definition_pandas",
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="redis_json_pandas_data_model",
             ),
@@ -64,8 +69,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "redis",
                 "redis_hashset_array_data_model",
                 {"collection_type": RedisCollectionTypes.HASHSET},
-                TestDataModelArray,
+                "dataclass_vector_data_model_array",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_ARRAY,
                 id="redis_hashset_array_data_model",
             ),
@@ -73,8 +82,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "redis",
                 "redis_hashset_list_data_model",
                 {"collection_type": RedisCollectionTypes.HASHSET},
-                TestDataModelList,
+                "dataclass_vector_data_model",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="redis_hashset_list_data_model",
             ),
@@ -83,7 +96,11 @@ class TestVectorStore(VectorStoreTestBase):
                 "redis_hashset_pandas_data_model",
                 {"collection_type": RedisCollectionTypes.HASHSET},
                 pd.DataFrame,
-                PANDAS_RECORD_DEFINITION,
+                "data_model_definition_pandas",
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="redis_hashset_pandas_data_model",
             ),
@@ -93,8 +110,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "azure_ai_search",
                 "azure_ai_search_array_data_model",
                 {},
-                TestDataModelArray,
+                "dataclass_vector_data_model_array",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_ARRAY,
                 id="azure_ai_search_array_data_model",
             ),
@@ -102,8 +123,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "azure_ai_search",
                 "azure_ai_search_list_data_model",
                 {},
-                TestDataModelList,
+                "dataclass_vector_data_model",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="azure_ai_search_list_data_model",
             ),
@@ -112,7 +137,11 @@ class TestVectorStore(VectorStoreTestBase):
                 "azure_ai_search_pandas_data_model",
                 {},
                 pd.DataFrame,
-                PANDAS_RECORD_DEFINITION,
+                "data_model_definition_pandas",
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="azure_ai_search_pandas_data_model",
             ),
@@ -122,8 +151,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "qdrant",
                 "qdrant_array_data_model",
                 {},
-                TestDataModelArray,
+                "dataclass_vector_data_model_array",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_ARRAY,
                 id="qdrant_array_data_model",
             ),
@@ -131,8 +164,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "qdrant",
                 "qdrant_list_data_model",
                 {},
-                TestDataModelList,
+                "dataclass_vector_data_model",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="qdrant_list_data_model",
             ),
@@ -141,7 +178,11 @@ class TestVectorStore(VectorStoreTestBase):
                 "qdrant_pandas_data_model",
                 {},
                 pd.DataFrame,
-                PANDAS_RECORD_DEFINITION,
+                "data_model_definition_pandas",
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="qdrant_pandas_data_model",
             ),
@@ -149,8 +190,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "qdrant_in_memory",
                 "qdrant_in_memory_array_data_model",
                 {},
-                TestDataModelArray,
+                "dataclass_vector_data_model_array",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_ARRAY,
                 id="qdrant_in_memory_array_data_model",
             ),
@@ -158,8 +203,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "qdrant_in_memory",
                 "qdrant_in_memory_list_data_model",
                 {},
-                TestDataModelList,
+                "dataclass_vector_data_model",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="qdrant_in_memory_list_data_model",
             ),
@@ -168,7 +217,11 @@ class TestVectorStore(VectorStoreTestBase):
                 "qdrant_in_memory_pandas_data_model",
                 {},
                 pd.DataFrame,
-                PANDAS_RECORD_DEFINITION,
+                "data_model_definition_pandas",
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="qdrant_in_memory_pandas_data_model",
             ),
@@ -176,8 +229,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "qdrant",
                 "qdrant_grpc_array_data_model",
                 {"prefer_grpc": True},
-                TestDataModelArray,
+                "dataclass_vector_data_model_array",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_ARRAY,
                 id="qdrant_grpc_array_data_model",
             ),
@@ -185,8 +242,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "qdrant",
                 "qdrant_grpc_list_data_model",
                 {"prefer_grpc": True},
-                TestDataModelList,
+                "dataclass_vector_data_model",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="qdrant_grpc_list_data_model",
             ),
@@ -195,7 +256,11 @@ class TestVectorStore(VectorStoreTestBase):
                 "qdrant_grpc_pandas_data_model",
                 {"prefer_grpc": True},
                 pd.DataFrame,
-                PANDAS_RECORD_DEFINITION,
+                "data_model_definition_pandas",
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="qdrant_grpc_pandas_data_model",
             ),
@@ -205,8 +270,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "weaviate_local",
                 "weaviate_local_array_data_model",
                 {},
-                TestDataModelArray,
+                "dataclass_vector_data_model_array",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_ARRAY,
                 id="weaviate_local_array_data_model",
             ),
@@ -214,8 +283,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "weaviate_local",
                 "weaviate_local_list_data_model",
                 {},
-                TestDataModelList,
+                "dataclass_vector_data_model",
                 None,
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="weaviate_local_list_data_model",
             ),
@@ -224,7 +297,11 @@ class TestVectorStore(VectorStoreTestBase):
                 "weaviate_local_pandas_data_model",
                 {},
                 pd.DataFrame,
-                PANDAS_RECORD_DEFINITION,
+                "data_model_definition_pandas",
+                None,
+                None,
+                None,
+                5,
                 RAW_RECORD_LIST,
                 id="weaviate_local_pandas_data_model",
             ),
@@ -234,8 +311,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "azure_cosmos_db_no_sql",
                 "azure_cosmos_db_no_sql_array_data_model",
                 {},
-                TestDataModelArrayFlat,
+                "dataclass_vector_data_model_array",
                 None,
+                None,
+                "flat",
+                None,
+                5,
                 RAW_RECORD_ARRAY,
                 marks=pytest.mark.skipif(
                     platform.system() != "Windows",
@@ -247,8 +328,12 @@ class TestVectorStore(VectorStoreTestBase):
                 "azure_cosmos_db_no_sql",
                 "azure_cosmos_db_no_sql_list_data_model",
                 {},
-                TestDataModelListFlat,
+                "dataclass_vector_data_model",
                 None,
+                None,
+                "flat",
+                None,
+                5,
                 RAW_RECORD_LIST,
                 marks=pytest.mark.skipif(
                     platform.system() != "Windows",
@@ -261,7 +346,11 @@ class TestVectorStore(VectorStoreTestBase):
                 "azure_cosmos_db_no_sql_pandas_data_model",
                 {},
                 pd.DataFrame,
-                PANDAS_RECORD_DEFINITION_FLAT,
+                "data_model_definition_pandas",
+                None,
+                "flat",
+                None,
+                5,
                 RAW_RECORD_LIST,
                 marks=pytest.mark.skipif(
                     platform.system() != "Windows",
@@ -278,10 +367,20 @@ class TestVectorStore(VectorStoreTestBase):
         store_id: str,
         collection_name: str,
         collection_options: dict[str, Any],
-        data_model_type: type,
-        data_model_definition: VectorStoreRecordDefinition | None,
+        data_model_type: str | Type,
+        data_model_definition: str | None,
+        distance_function,
+        index_kind,
+        vector_property_type,
+        dimensions,
         record: dict[str, Any],
+        request,
     ):
+        """Test vector store functionality."""
+        if isinstance(data_model_type, str):
+            data_model_type = request.getfixturevalue(data_model_type)
+        if data_model_definition is not None:
+            data_model_definition = request.getfixturevalue(data_model_definition)
         try:
             async with (
                 stores[store_id] as vector_store,
@@ -290,7 +389,8 @@ class TestVectorStore(VectorStoreTestBase):
                 ) as collection,
             ):
                 try:
-                    await collection.create_collection_if_not_exists()
+                    await collection.delete_collection()
+                    await collection.create_collection()
                 except Exception as exc:
                     pytest.fail(f"Failed to create collection: {exc}")
